osha-chatbot-v2/requirements_fixed.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
arize-otel==0.8.0
arize-phoenix==8.15.0
arize-phoenix-client==1.1.0
arize-phoenix-evals==0.20.3
arize-phoenix-otel==0.8.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.5.1
azure-core==1.32.0
azure-identity==1.21.0
beautifulsoup4==4.13.3
boto3==1.37.16
botocore==1.37.16
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dacite==1.9.2
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
guardrails-ai==0.6.5
googleapis-common-protos==1.69.2
graphql-core==3.2.6
greenlet==3.1.1
grpc-interceptor==0.15.4
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
llama-cloud==0.1.15
llama-cloud-services==0.6.6
llama-index==0.12.25
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.25
llama-index-embeddings-azure-openai==0.3.1
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-azure-openai==0.3.2
llama-index-llms-openai==0.3.26
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.2.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.68.0
openinference-instrumentation==0.1.24
openinference-instrumentation-bedrock==0.1.17
openinference-instrumentation-llama-index==3.3.3
openinference-instrumentation-openai==0.1.22
openinference-semantic-conventions==0.1.15
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.39
sqlean.py==3.47.0
starlette==0.46.1
strawberry-graphql==0.262.5
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
